pointerevents/touch-action-inherit_child-pan-x-child-pan-x_touch.html

<!doctype html>
<html>
	<head>
		<title>CSS</title>
		<meta name="viewport" content="width=device-width">
		<script src="http://w3c-test.org/resources/testharness.js"></script>
		<script src="pe_functions.js"></script>
		<style>
			.scroller {
				width: 700px;
				height: 430px;
				margin: 20px;
				overflow: auto;
				background: black;
			}
			.scroller > div {
				height: 1000px;
				width: 1000px;
				color: white;
			}
			.scroller > div div {
				height: 100%;
				width: 100%;
				color: white;
			}		
		</style>
	</head>
	<body onload="run()">
		<h2>Pointer Events touch-action attribute support</h2>
		<h4>Test Description: using touch, slowly perform an interaction that would normally trigger a pan in the x and y axes.</h4>
		<p>Note: this test is for touch-devices only.</p>
		<div class="scroller" id="target0">
			<div style="touch-action: pan-x">
				<div style="touch-action: pan-x"> Expected: only pans in x direction </div>
			</div>
		</div>

		<script type='text/javascript'>
		var detected_pointertypes = {};

		var xCount = 0;
		var yCount = 0;
		var countToPass = 50;
		var xScr0, yScr0, xScr1, yScr1;

		setup({ explicit_done: true });
		add_completion_callback(showPointerTypes);

		function run() {
			var target0 = document.getElementById("target0");
			
			var test_touchaction = async_test("touch-action attribute test");
			
			xScr0 = target0.scrollLeft;
			yScr0 = target0.scrollTop;
			
			on_event(target0, 'pointerdown', function(event) {
				detected_pointertypes[event.pointerType] = true;
			})
			
			target0.addEventListener('scroll', function(event) {
				xScr1 = target0.scrollLeft;
				yScr1 = target0.scrollTop;
				
				if(xScr1 != xScr0) {
					xCount++;
					xScr0 = xScr1;
				}
				
				if(yScr1 != yScr0) {
					test(function() {
						assert_true(yScr1 == yScr0,
						"y-scroll received while pan-x is set");
					}, "y-scroll received while pan-x is set");
					yCount++;
				}
					
				if(xCount >= countToPass && yCount == 0) {
					test_touchaction.done();
				}
			});
		}
		</script>
		<style>
			div {
				margin: 0em;
				padding: 2em;
			}
			#complete-notice {
				background: #afa;
				border: 1px solid #0a0;
				display: none;
			}
			#pointertype-log {
				font-weight: bold;
			}
		</style>
		<div id="complete-notice">
			<p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
		</div>
		<div id="log"></div>
	</body>
</html>
